+++ tomcat/src/main/java/nextstep/jwp/controller/LoginController.java
@@ -3,15 +3,16 @@ package nextstep.jwp.controller;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import nextstep.jwp.controller.FileReader.FileReader;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import nextstep.jwp.session.HttpSessionManager;
 import org.apache.catalina.session.Session;
+import org.apache.coyote.AbstractController;
+import org.apache.coyote.fileReader.FileReader;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.request.Request;
 import org.apache.coyote.response.HttpStatus;
 import org.apache.coyote.response.Response;
-import org.apache.coyote.session.HttpSessionManager;
 
 public class LoginController extends AbstractController {
