+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -4,7 +4,8 @@ import java.util.Map;
 import java.util.Set;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import nextstep.jwp.controller.FileReader.FileReader;
+import org.apache.coyote.fileReader.FileReader;
+import org.apache.coyote.AbstractController;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.request.Request;
 import org.apache.coyote.response.HttpStatus;
