+++ tomcat/src/main/java/org/apache/coyote/exception/UncheckedServletException.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.exception;
+package org.apache.coyote.exception;
 
 public class UncheckedServletException extends RuntimeException {
