+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -4,8 +4,8 @@ import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.IOException;
 import java.net.Socket;
-import nextstep.jwp.exception.UncheckedServletException;
-import nextstep.jwp.controller.Controller;
+import org.apache.coyote.exception.UncheckedServletException;
+import org.apache.coyote.Controller;
 import org.apache.coyote.Processor;
 import org.apache.coyote.request.Request;
 import org.apache.coyote.request.RequestParser;
